+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 Valdi Changelog
 ===============
 
-## 0.12.0
+## 1.0.0
 
 Released: Upcoming
 
+## 0.12.0
+
+Released: 2019-09-09
+
 - Attention: Renamed the RuleBuilder function "addFieldRule" to "field"
 - Attention: Renamed the RuleBuilder function "addRule" to "rule"
+- Added examples to all validators to the manual
   
 ## 0.11.0